backend/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-GoogleStorage==0.1.2
Flask-PyMongo==2.3.0
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.0
Pillow==8.1.2
pipenv==2022.10.12
platformdirs==2.5.2
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
tenacity==8.1.0
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
zipp==3.9.0
